requirements.txt
attrs                       25.3.0
azure-core                  1.35.0
azure-data-tables           12.7.0
azure-identity              1.23.0
azure-keyvault              4.2.0
azure-keyvault-certificates 4.10.0
azure-keyvault-keys         4.11.0
azure-keyvault-secrets      4.10.0
azure-storage-blob          12.25.1
babel                       2.17.0
behave                      1.2.6
behavex                     4.1.0
behavex-images              3.2.1
certifi                     2025.6.15        
cffi                        1.17.1
charset-normalizer          3.4.2
configobj                   5.0.9
cryptography                45.0.5
csscompressor               0.9.5
customtkinter               5.2.2
darkdetect                  0.8.0
docxcompose                 1.4.0
docxtpl                     0.20.0
et_xmlfile                  2.0.0
filelock                    3.18.0
h11                         0.16.0
htmlmin                     0.1.12
idna                        3.10
isodate                     0.7.2
Jinja2                      3.1.6
lxml                        6.0.0
MarkupSafe                  3.0.2
msal                        1.32.3
msal-extensions             1.3.1
multidict                   6.6.3
openpyxl                    3.1.5
outcome                     1.3.0.post0      
packaging                   25.0
parse                       1.20.2
parse_type                  0.6.4
pillow                      11.3.0
pip                         25.1.1
propcache                   0.3.2
pycparser                   2.22
PyJWT                       2.10.1
PySocks                     1.7.1
python-docx                 1.2.0
requests                    2.32.4
selenium                    4.34.0
setuptools                  80.9.0
six                         1.17.0
sniffio                     1.3.1
sortedcontainers            2.4.0
trio                        0.30.0
trio-websocket              0.12.2
typing_extensions           4.14.0
urllib3                     2.4.0
websocket-client            1.8.0
wsproto                     1.2.0
yarl                        1.20.1